refactor: rename usePageTitle to useHeaderTitle and split read/write
Splits the hook into useHeaderTitle() (read) and useDeclareHeaderTitle(title)
(write), so pages declare their title as a hook argument instead of running a
useEffect with a setter. Removes the per-page effect boilerplate, the
ownsTitle ref, and the document.title vs toolbar-title naming ambiguity.
