+++ src/pages/BoardPage.tsx
@@ -1,18 +1,14 @@
-import { usePageTitle } from "@app/usePageTitle";
+import { useDeclareHeaderTitle } from "@app/useHeaderTitle";
 import { BoardView, useBoard, type BoardRouteParams } from "@features/board";
 import { ErrorFallback } from "@shared/components/ErrorFallback";
 import { LoadingIndicator } from "@shared/components/LoadingIndicator";
-import { useEffect } from "react";
 import { useParams } from "react-router";
 
 function BoardPage() {
   const { setId = "", boardId = "" } = useParams<BoardRouteParams>();
   const { board, error } = useBoard({ setId, boardId });
-  const { setPageTitle } = usePageTitle();
 
-  useEffect(() => {
-    setPageTitle(board?.name);
-  }, [setPageTitle, board?.name]);
+  useDeclareHeaderTitle(board?.name);
 
   if (error) {
     return (
